got flip working hard split between game engine and love framework
