+++ carts/bench/conf.lua
@@ -1,0 +1,4 @@
+function love.conf(t)
+    t.console = true
+    t.window = false
+end
