+++ carts/game/engine/conf.lua
@@ -1,0 +1,11 @@
+local conf={}
+conf.start_level="e1m1"
+-- game key bindings
+conf.keys = {
+    up = {'w','up'},
+    down = {'s','down'},
+    left = {'a','left'},
+    right = {'d','right'},
+    action = {'space'}
+}
+return conf
