--- carts/game/menu.lua
@@ -1,18 +1,0 @@
-local ui={}
-
-function ui:window(args)
-end
-
-function ui:list(args)
-end
-
-function ui:menu(args)
-end
-
-function ui:update()
-end
-
-function ui:render()
-end
-
-return ui
