+++ carts/game/platforms/love/input.lua
@@ -1,0 +1,6 @@
+local Input={}
+local lk=love.input
+function Input.isScancodeDown(k)
+    return lk.isScancodeDown(k)
+end
+return Input
